Renames as BroadcastDetail
